Правки и изменения согласно замечаниям.
