+++ index.html
@@ -120,10 +120,11 @@
                   <div class="img-upload__field-wrapper">
                     <input class="text__hashtags" name="hashtags" id="hashtags" placeholder="#ХэшТег"
                     maxlength="104">
+
                   </div>
                   <div class="img-upload__field-wrapper">
                     <textarea class="text__description" name="description" placeholder="Ваш комментарий..."
-                    id = "commentfield" maxlength="140">
+                    id = "commentfield" >
                     </textarea>
                   </div>
                 </fieldset>
